Fix player in separate file
